renderer/c: encoder skeleton — viewport walk + base flags + colors
Implements the inner loop of encodeCellsPhase1: PageList iteration
over the visible viewport, writing fg/bg colors and cell-flag-derived
bits into the packed output buffer. No atlas, no decorations, no
special cells yet — those land in subsequent commits.

Host tests cover empty grid (default-empty flags) and single ASCII
bold cell (FLAG_BOLD + FLAG_USE_THEME_BG).

Ancillary plumbing required to make this work:

- terminal/c/terminal.zig: expose terminalPtr() so the renderer C ABI
  can recover the inner Terminal from the opaque wrapper handle the
  TS side passes through FrameCtx.terminal_handle.

- renderer/c/frame_ctx.zig: widen the pointer-typed fields from u32 to
  usize. wasm32's usize is still u32 (no ABI change for JS), but the
  host Zig tests run on a 64-bit target and need real pointers in
  terminal_handle / output_buf_ptr / etc.

- renderer/c/encode_cells.zig: re-declare cell-layout constants
  locally instead of importing ../cell.zig — the latter pulls in
  renderer.zig + font/main.zig with their Metal/OpenGL/font-asset
  dependencies, none of which compile in the vt-only test target.
  A comptime drift check in renderer/c/main.zig pins the local
  constants against the canonical ones so they can't silently
  diverge.

- lib_vt.zig: hoist renderer/c/encode_cells.zig + frame_ctx.zig into
  the test root so their tests actually run (the previous Task-4 stub
  test was reachable only from the comptime export block, which is
  not visited by Zig's test discovery).

+++ src/renderer/c/frame_ctx.zig
@@ -4,12 +4,18 @@
 //! and the JS-side wrapper in lib/renderer-wasm-encode.ts. The layouts
 //! must match exactly — JS uses a DataView with byte offsets derived
 //! from these field orderings.
+//!
+//! Pointer-typed fields (terminal_handle, output_buf_ptr, etc.) use
+//! `usize` so the same struct compiles for both wasm32 (where usize is
+//! u32) and the 64-bit host used by Zig unit tests. On WASM the layout
+//! matches what the JS side writes; on the host the struct grows but
+//! never crosses an ABI boundary, so the discrepancy is harmless.
 
 /// Per-frame input written by JS into linear memory and read by
-/// renderer_encode_cells_phase1. Fields are u32 except for signed
-/// row/col coords that may be -1 (sentinel for "not set").
+/// renderer_encode_cells_phase1. Most fields are u32; pointer-typed
+/// fields are `usize` for wasm32/host-native portability.
 pub const FrameCtx = extern struct {
-    terminal_handle: u32,
+    terminal_handle: usize,
     viewport_y: u32,
     scrollback_len: u32,
 
@@ -45,22 +51,22 @@ pub const FrameCtx = extern struct {
     max_kitty_images: u32,
 
     /// Output: packed cell bytes. Sized for rows*cols*CELL_BYTES.
-    output_buf_ptr: u32,
+    output_buf_ptr: usize,
     output_buf_len: u32,
 
     /// Output: NeedsAtlasEntry array. Capacity = rows*cols.
-    needs_atlas_ptr: u32,
+    needs_atlas_ptr: usize,
     needs_atlas_capacity: u32,
 
     /// Output: grapheme UTF-8 scratch arena. Capacity should be
     /// rows*cols*32 bytes (generous upper bound).
-    grapheme_scratch_ptr: u32,
+    grapheme_scratch_ptr: usize,
     grapheme_scratch_len: u32,
 
     /// Input: JS-prepared imageId -> (idx in used_kitty_image_ids) map,
     /// for kitty placeholder cells. Format: pairs of (imageId, idx) u32,
     /// terminated by imageId==0. Read but not written by WASM.
-    kitty_image_table_ptr: u32,
+    kitty_image_table_ptr: usize,
     kitty_image_table_len: u32,
 };
